+++ contracts/interfaces/NativeSwap.sol
@@ -2,6 +2,21 @@
 pragma solidity ^0.6.12;
 
 interface NativeSwap {
+    struct PublicClaimInfo {
+        bytes32 preimage;
+        uint amount;
+        address claimAddress;
+        address refundAddress;
+        uint timelock;
+    }
+
+    struct ExternalClaimInfo {
+        bytes32 preimage;
+        uint amount;
+        address refundAddress;
+        uint timelock;
+    }
+
     function swaps(bytes32) external returns (bool);
 
     function hashValues(
