+++ contracts/verifier/MinimalBoltzDeployVerifier.sol
@@ -10,7 +10,8 @@ import "../factory/MinimalBoltzSmartWalletFactory.sol";
 import "../interfaces/IDeployVerifier.sol";
 import "../interfaces/EnvelopingTypes.sol";
 import "../utils/ContractValidator.sol";
-import "../utils/BoltzValidator.sol";
+import "../utils/BoltzBytesUtil.sol";
+import "../interfaces/NativeSwap.sol";
 
 /**
  * A Verifier to be used on deploys.
@@ -52,7 +53,7 @@ contract MinimalBoltzDeployVerifier is
 
         destinationContractValidation(relayRequest.request.to);
 
-        uint256 amount = BoltzValidator.validate(
+        uint256 amount = _validateClaim(
             relayRequest.request.data,
             relayRequest.request.to,
             contractAddr
@@ -76,5 +77,48 @@ contract MinimalBoltzDeployVerifier is
                 relayRequest.request.tokenContract
             )
         );
+    }
+
+    function _validateClaim(
+        bytes calldata data,
+        address to,
+        address contractAddr
+    ) private returns (uint256) {
+        bytes4 signature = BoltzBytesUtil.toBytes4(data, 0);
+        NativeSwap.PublicClaimInfo memory claim;
+
+        if (signature == BoltzBytesUtil._EXTERNAL_SIGNATURE) {
+            NativeSwap.ExternalClaimInfo memory localClaim = abi.decode(
+                data[4:],
+                (NativeSwap.ExternalClaimInfo)
+            );
+            claim = NativeSwap.PublicClaimInfo(
+                localClaim.preimage,
+                localClaim.amount,
+                contractAddr,
+                localClaim.refundAddress,
+                localClaim.timelock
+            );
+        } else if (signature == BoltzBytesUtil._PUBLIC_SIGNATURE) {
+            claim = abi.decode(data[4:], (NativeSwap.PublicClaimInfo));
+        } else {
+            revert("Signature not allowed");
+        }
+
+        NativeSwap swap = NativeSwap(to);
+
+        bytes32 preimageHash = sha256(abi.encodePacked(claim.preimage));
+
+        bytes32 hashValue = swap.hashValues(
+            preimageHash,
+            claim.amount,
+            claim.claimAddress,
+            claim.refundAddress,
+            claim.timelock
+        );
+
+        require(swap.swaps(hashValue), "Verifier: swap has no RBTC");
+
+        return claim.amount;
     }
 }
