+++ test/smartwallet/boltzSmartWallet.test.ts
@@ -3,14 +3,13 @@ import { BaseProvider } from '@ethersproject/providers';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import chai, { expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import { Wallet, constants, ethers } from 'ethers';
+import { Wallet } from 'ethers';
 import { ethers as hardhat } from 'hardhat';
 import {
   ERC20,
   BoltzSmartWallet,
   BoltzSmartWalletFactory,
   BoltzSmartWallet__factory,
-  NativeSwap,
 } from 'typechain-types';
 import { createValidPersonalSignSignature } from '../utils/createValidPersonalSignSignature';
 import {
@@ -764,7 +763,7 @@ describe('BoltzSmartWallet contract', function () {
     let mockSmartWallet: MockContract<BoltzSmartWallet>;
     let provider: BaseProvider;
     let owner: Wallet;
-    let recipient: FakeContract<NativeSwap>;
+    let recipient: FakeContract<BoltzSmartWallet>;
     let recipientFunction: string;
     let privateKey: Buffer;
     let worker: SignerWithAddress;
@@ -798,19 +797,12 @@ describe('BoltzSmartWallet contract', function () {
         value: hardhat.utils.parseEther('1'),
       });
 
-      recipient = await smock.fake('NativeSwap');
-      recipient.swaps.returns(true);
+      recipient = await smock.fake('BoltzSmartWallet');
+      recipient.isInitialized.returns(true);
 
-      const ABI = [
-        'function claim(bytes32 preimage, uint amount, address refundAddress, uint timelock)',
-      ];
+      const ABI = ['function isInitialized()'];
       const abiInterface = new hardhat.utils.Interface(ABI);
-      recipientFunction = abiInterface.encodeFunctionData('claim', [
-        constants.HashZero,
-        ethers.utils.parseEther('0.5'),
-        constants.AddressZero,
-        500,
-      ]);
+      recipientFunction = abiInterface.encodeFunctionData('isInitialized', []);
 
       privateKey = Buffer.from(owner.privateKey.substring(2, 66), 'hex');
 
@@ -852,6 +844,8 @@ describe('BoltzSmartWallet contract', function () {
       ).not.to.be.rejected;
 
       expect(fakeToken.transfer, 'Token.transfer was called').not.to.be.called;
+      expect(recipient.isInitialized, 'Recipient method was not called').to.be
+        .called;
     });
 
     it('Should pay for relay using token', async function () {
@@ -888,6 +882,8 @@ describe('BoltzSmartWallet contract', function () {
       ).not.to.be.rejected;
 
       expect(fakeToken.transfer, 'Token.transfer() was not called').to.be
+        .called;
+      expect(recipient.isInitialized, 'Recipient method was not called').to.be
         .called;
     });
 
@@ -942,6 +938,8 @@ describe('BoltzSmartWallet contract', function () {
       expect(finalWorkerBalance).to.be.equal(
         initialWorkerBalance.add(amountToBePaid)
       );
+      expect(recipient.isInitialized, 'Recipient method was not called').to.be
+        .called;
     });
 
     it('Should increment nonce', async function () {
@@ -1056,8 +1054,6 @@ describe('BoltzSmartWallet contract', function () {
         relayHub: relayHub.address,
         from: owner.address,
         validUntilTime: expirationInSeconds, //Always 1 day (86400 sec) ahead
-        data: recipientFunction,
-        to: recipient.address,
       });
 
       const typedRequestData = new TypedRequestData(
