+++ test/verifier/minimalBoltzDeployVerifier.test.ts
@@ -567,56 +567,190 @@ describe('MinimalBoltzDeployVerifier Contract', function () {
       });
     });
 
-    describe('external method', function () {
-      let data: string;
+    describe('calling claim', function () {
+      type ClaimMethodType = 'public' | 'external';
 
-      beforeEach(function () {
-        fakeSwap.swaps.returns(true);
-        const ABI = [
-          'function claim(bytes32 preimage, uint amount, address refundAddress, uint timelock)',
-        ];
-        const abiInterface = new ethers.utils.Interface(ABI);
-        data = abiInterface.encodeFunctionData('claim', [
-          constants.HashZero,
-          ethers.utils.parseEther('0.5'),
-          constants.AddressZero,
-          500,
-        ]);
-      });
+      function testClaimMethodType(method: ClaimMethodType) {
+        describe(`${method} method`, function () {
+          let data: string;
+          let smartWalletAddress: string;
 
-      it('should not revert if swap contract can pay for the native fee', async function () {
-        const smartWalletAddress = ethers.Wallet.createRandom().address;
-        fakeWalletFactory.getSmartWalletAddress.returns(smartWalletAddress);
+          beforeEach(function () {
+            fakeSwap.swaps.returns(true);
+            const ABI = [
+              'function claim(bytes32 preimage, uint amount, address claimAddress, address refundAddress, uint timelock)',
+              'function claim(bytes32 preimage, uint amount, address refundAddress, uint timelock)',
+            ];
+            const abiInterface = new ethers.utils.Interface(ABI);
+            smartWalletAddress = ethers.Wallet.createRandom().address;
+            switch (method) {
+              case 'external':
+                data = abiInterface.encodeFunctionData(
+                  'claim(bytes32,uint256,address,uint256)',
+                  [
+                    constants.HashZero,
+                    ethers.utils.parseEther('0.5'),
+                    smartWalletAddress,
+                    500,
+                  ]
+                );
+                break;
+              case 'public':
+                data = abiInterface.encodeFunctionData(
+                  'claim(bytes32,uint256,address,address,uint256)',
+                  [
+                    constants.HashZero,
+                    ethers.utils.parseEther('0.5'),
+                    smartWalletAddress,
+                    constants.AddressZero,
+                    500,
+                  ]
+                );
+                break;
+            }
+          });
 
-        const deployRequest: EnvelopingTypes.DeployRequestStruct = {
-          relayData: {
-            callForwarder: fakeWalletFactory.address,
-            callVerifier: deployVerifierMock.address,
-            gasPrice: '10',
-            feesReceiver: relayWorker.address,
-          },
-          request: {
-            recoverer: constants.AddressZero,
-            index: '0',
-            data,
-            from: owner.address,
-            to: fakeSwap.address,
-            nonce: '0',
-            tokenGas: '50000',
-            relayHub: fakeRelayHub.address,
-            tokenAmount: ethers.utils.parseEther('0.5'),
-            tokenContract: constants.AddressZero,
-            validUntilTime: '0',
-            value: '0',
-          },
-        };
+          it('should revert if native balance is too low', async function () {
+            const deployRequest: EnvelopingTypes.DeployRequestStruct = {
+              relayData: {
+                callForwarder: fakeWalletFactory.address,
+                callVerifier: deployVerifierMock.address,
+                gasPrice: '10',
+                feesReceiver: relayWorker.address,
+              },
+              request: {
+                recoverer: constants.AddressZero,
+                index: '0',
+                data,
+                from: owner.address,
+                to: fakeSwap.address,
+                nonce: '0',
+                tokenGas: '50000',
+                relayHub: fakeRelayHub.address,
+                tokenAmount: ethers.utils.parseEther('1'),
+                tokenContract: constants.AddressZero,
+                validUntilTime: '0',
+                value: '0',
+              },
+            };
 
-        const result = deployVerifierMock.verifyRelayedCall(
-          deployRequest,
-          '0x00'
-        );
-        await expect(result).to.not.be.reverted;
-      });
+            const result = deployVerifierMock.verifyRelayedCall(
+              deployRequest,
+              '0x00'
+            );
+            await expect(result).to.be.revertedWith('Native balance too low');
+          });
+
+          it('should revert if swap contract cannot pay for the native fee', async function () {
+            fakeWalletFactory.getSmartWalletAddress.returns(smartWalletAddress);
+            fakeSwap.swaps.returns(false);
+
+            const deployRequest: EnvelopingTypes.DeployRequestStruct = {
+              relayData: {
+                callForwarder: fakeWalletFactory.address,
+                callVerifier: deployVerifierMock.address,
+                gasPrice: '10',
+                feesReceiver: relayWorker.address,
+              },
+              request: {
+                recoverer: constants.AddressZero,
+                index: '0',
+                data,
+                from: owner.address,
+                to: fakeSwap.address,
+                nonce: '0',
+                tokenGas: '50000',
+                relayHub: fakeRelayHub.address,
+                tokenAmount: ethers.utils.parseEther('0.5'),
+                tokenContract: constants.AddressZero,
+                validUntilTime: '0',
+                value: '0',
+              },
+            };
+
+            const result = deployVerifierMock.verifyRelayedCall(
+              deployRequest,
+              '0x00'
+            );
+            await expect(result).to.be.revertedWith(
+              'Verifier: swap has no RBTC'
+            );
+          });
+
+          it('should not revert if swap contract can pay for the native fee', async function () {
+            fakeWalletFactory.getSmartWalletAddress.returns(smartWalletAddress);
+
+            const deployRequest: EnvelopingTypes.DeployRequestStruct = {
+              relayData: {
+                callForwarder: fakeWalletFactory.address,
+                callVerifier: deployVerifierMock.address,
+                gasPrice: '10',
+                feesReceiver: relayWorker.address,
+              },
+              request: {
+                recoverer: constants.AddressZero,
+                index: '0',
+                data,
+                from: owner.address,
+                to: fakeSwap.address,
+                nonce: '0',
+                tokenGas: '50000',
+                relayHub: fakeRelayHub.address,
+                tokenAmount: ethers.utils.parseEther('0.5'),
+                tokenContract: constants.AddressZero,
+                validUntilTime: '0',
+                value: '0',
+              },
+            };
+
+            const result = deployVerifierMock.verifyRelayedCall(
+              deployRequest,
+              '0x00'
+            );
+            await expect(result).to.not.be.reverted;
+          });
+
+          it('should not revert if not paying', async function () {
+            const deployRequest: EnvelopingTypes.DeployRequestStruct = {
+              relayData: {
+                callForwarder: fakeWalletFactory.address,
+                callVerifier: deployVerifierMock.address,
+                gasPrice: '10',
+                feesReceiver: relayWorker.address,
+              },
+              request: {
+                recoverer: constants.AddressZero,
+                index: '0',
+                data,
+                from: owner.address,
+                to: fakeSwap.address,
+                nonce: '0',
+                tokenGas: '50000',
+                relayHub: fakeRelayHub.address,
+                tokenAmount: '0',
+                tokenContract: constants.AddressZero,
+                validUntilTime: '0',
+                value: '0',
+              },
+            };
+
+            const result = deployVerifierMock.verifyRelayedCall(
+              deployRequest,
+              '0x00'
+            );
+
+            await expect(result).to.not.be.reverted;
+          });
+        });
+      }
+
+      // Using [dynamically generated tests](https://mochajs.org/#dynamically-generating-tests)
+      // we needs the mocha/no-setup-in-describe rule to be disabled
+      // see: https://github.com/lo1tuma/eslint-plugin-mocha/blob/main/docs/rules/no-setup-in-describe.md#disallow-setup-in-describe-blocks-mochano-setup-in-describe
+      // eslint-disable-next-line mocha/no-setup-in-describe
+      testClaimMethodType('public');
+      // eslint-disable-next-line mocha/no-setup-in-describe
+      testClaimMethodType('external');
     });
   });
 });
